refactor: move domain from src/server to src/
