+++ src/server/handlers/consents/{ID}/generateChallenge.ts
@@ -28,7 +28,7 @@
  --------------
  ******/
 // eslint-disable-next-line max-len
-import { updateConsentCredential, isConsentRequestInitiatedByValidSource, putConsentId, ConsentCredential } from '../../../domain/consents/{ID}/generateChallenge'
+import { updateConsentCredential, isConsentRequestInitiatedByValidSource, putConsentId, ConsentCredential } from '../../../../domain/consents/{ID}/generateChallenge'
 import * as challenge from '../../../../lib/challenge'
 import { Request, ResponseToolkit, ResponseObject } from '@hapi/hapi'
 import { consentDB, scopeDB } from '../../../../lib/db'
